requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
rich
GitPython
nbformat
bandit
radon
tox